+++ src/features/album/components/AlbumSection.tsx
@@ -1,77 +1,28 @@
-import TrackListItem from "@/features/album/components/TrackListItem";
+import AlbumInfo from "@/features/album/components/AlbumInfo";
+import TrackList from "@/features/album/components/TrackList";
+import useAlbumPlayer from "@/features/album/hooks/useAlbumPlayer";
 import AudioPlayer from "@/features/audio/components/AudioPlayer";
 import LikeButton from "@/features/liked/components/LikeButton";
-import { SimplifiedAlbum, SimplifiedTrack } from "@/types/album";
-import { useTranslation } from "next-i18next";
+import { SimplifiedAlbum } from "@/types/album";
 import Image from "next/image";
-import Link from "next/link";
-import { useEffect, useState } from "react";
 
 interface AlbumSectionProps {
   album: SimplifiedAlbum;
 }
 
 const AlbumSection = ({ album }: AlbumSectionProps) => {
-  const { t } = useTranslation(["common", "album"]);
-
-  const [currentTrackIndex, setCurrentTrackIndex] = useState<number | null>(
-    null,
-  );
-  const [isPlaying, setIsPlaying] = useState<boolean>(false);
-  const [volume, setVolume] = useState<number>(() => {
-    if (typeof window !== "undefined") {
-      const savedVolume = localStorage.getItem("volume");
-      return savedVolume !== null ? Number(savedVolume) : 1;
-    }
-    return 1;
-  });
-
-  useEffect(() => {
-    const savedVolume = localStorage.getItem("volume");
-    if (savedVolume !== null) {
-      setVolume(Number(savedVolume));
-    }
-  }, []);
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      localStorage.setItem("volume", volume.toString());
-    }
-  }, [volume]);
-
-  const handlePlay = (index: number) => {
-    if (currentTrackIndex === index) {
-      setIsPlaying((prev) => !prev);
-    } else {
-      setCurrentTrackIndex(index);
-      setIsPlaying(true);
-    }
-  };
-
-  const handlePrevious = () => {
-    if (currentTrackIndex !== null && currentTrackIndex > 0) {
-      setCurrentTrackIndex(currentTrackIndex - 1);
-      setIsPlaying(true);
-    }
-  };
-
-  const handleNext = () => {
-    if (
-      currentTrackIndex !== null &&
-      currentTrackIndex < album.tracks.items.length - 1
-    ) {
-      setCurrentTrackIndex(currentTrackIndex + 1);
-      setIsPlaying(true);
-    }
-  };
-
-  const handleClosePlayer = () => {
-    setCurrentTrackIndex(null);
-    setIsPlaying(false);
-  };
-
-  const currentTrack =
-    currentTrackIndex !== null ? album.tracks.items[currentTrackIndex] : null;
+  const {
+    currentTrackIndex,
+    isPlaying,
+    volume,
+    setVolume,
+    setIsPlaying,
+    handlePlay,
+    handlePrevious,
+    handleNext,
+    handleClosePlayer,
+    currentTrack,
+  } = useAlbumPlayer(album);
 
   return (
     <div
@@ -95,49 +46,15 @@ const AlbumSection = ({ album }: AlbumSectionProps) => {
           releaseDate={album.releaseDate}
         />
       </div>
-      <div className="w-full max-w-4xl">
-        <p className="text-gray-400 mt-2">
-          {t("artists", { ns: "album" })}:{" "}
-          {album.artists.map((artist, index) => (
-            <span key={artist.id} className="inline-flex items-center">
-              <Link href={`/artist/${artist.id}`}>
-                <span className="text-chefchaouenBlue hover:underline">
-                  {artist.name}
-                </span>
-              </Link>
-              {index < album.artists.length - 1 && (
-                <span className="ml-1 mr-2">,</span>
-              )}
-            </span>
-          ))}
-        </p>
-        <p className="text-gray-400 mt-1">
-          {t("release_date", { ns: "album" })}: {album.releaseDate}
-        </p>
-        <p className="text-gray-400 mt-1">
-          {t("total_tracks", { ns: "album" })}: {album.totalTracks}
-        </p>
-        <p className="text-gray-400 mt-1">
-          {t("label", { ns: "album" })}: {album.label}
-        </p>
-      </div>
-      <div className="w-full mt-4">
-        <h2 className="text-2xl font-semibold text-white mb-2">
-          {t("tracks", { ns: "album" })}
-        </h2>
-        <ul className="w-full space-y-4">
-          {album.tracks.items.map((track: SimplifiedTrack, index: number) => (
-            <TrackListItem
-              key={track.id}
-              index={index}
-              track={track}
-              onPlay={() => handlePlay(index)}
-              isCurrent={currentTrackIndex === index}
-              isPlaying={isPlaying}
-            />
-          ))}
-        </ul>
-      </div>
+
+      <AlbumInfo album={album} />
+      <TrackList
+        tracks={album.tracks.items}
+        currentTrackIndex={currentTrackIndex}
+        isPlaying={isPlaying}
+        onPlay={handlePlay}
+      />
+
       <AudioPlayer
         track={currentTrack}
         isPlaying={isPlaying}
