+++ src/features/album/hooks/useAlbumPlayer.ts
@@ -1,0 +1,78 @@
+import { SimplifiedAlbum } from "@/types/album";
+import { useEffect, useState } from "react";
+
+const useAlbumPlayer = (album: SimplifiedAlbum) => {
+  const [currentTrackIndex, setCurrentTrackIndex] = useState<number | null>(
+    null,
+  );
+  const [isPlaying, setIsPlaying] = useState<boolean>(false);
+  const [volume, setVolume] = useState<number>(() => {
+    if (typeof window !== "undefined") {
+      const savedVolume = localStorage.getItem("volume");
+      return savedVolume !== null ? Number(savedVolume) : 1;
+    }
+    return 1;
+  });
+
+  useEffect(() => {
+    const savedVolume = localStorage.getItem("volume");
+    if (savedVolume !== null) {
+      setVolume(Number(savedVolume));
+    }
+  }, []);
+
+  useEffect(() => {
+    if (typeof window !== "undefined") {
+      localStorage.setItem("volume", volume.toString());
+    }
+  }, [volume]);
+
+  const handlePlay = (index: number) => {
+    if (currentTrackIndex === index) {
+      setIsPlaying((prev) => !prev);
+    } else {
+      setCurrentTrackIndex(index);
+      setIsPlaying(true);
+    }
+  };
+
+  const handlePrevious = () => {
+    if (currentTrackIndex !== null && currentTrackIndex > 0) {
+      setCurrentTrackIndex(currentTrackIndex - 1);
+      setIsPlaying(true);
+    }
+  };
+
+  const handleNext = () => {
+    if (
+      currentTrackIndex !== null &&
+      currentTrackIndex < album.tracks.items.length - 1
+    ) {
+      setCurrentTrackIndex(currentTrackIndex + 1);
+      setIsPlaying(true);
+    }
+  };
+
+  const handleClosePlayer = () => {
+    setCurrentTrackIndex(null);
+    setIsPlaying(false);
+  };
+
+  const currentTrack =
+    currentTrackIndex !== null ? album.tracks.items[currentTrackIndex] : null;
+
+  return {
+    currentTrackIndex,
+    isPlaying,
+    volume,
+    setVolume,
+    setIsPlaying,
+    handlePlay,
+    handlePrevious,
+    handleNext,
+    handleClosePlayer,
+    currentTrack,
+  };
+};
+
+export default useAlbumPlayer;
