+++ src/pages/album/[albumId].tsx
@@ -1,7 +1,9 @@
 import AlbumSection from "@/features/album/components/AlbumSection";
+import AlbumSEO from "@/features/album/components/AlbumSEO";
 import useFetchAlbum from "@/features/album/queries/useFetchAlbum";
 import ErrorComponent from "@/features/common/components/ErrorComponent";
 import LoadingBar from "@/features/common/components/LoadingBar";
+import handlePageError from "@/libs/utils/handlePageError";
 import { GetServerSideProps } from "next";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
@@ -12,7 +14,6 @@ interface AlbumPageProps {
 
 const AlbumPage = ({ albumId }: AlbumPageProps) => {
   const { t } = useTranslation(["common", "album", "error"]);
-
   const { data, error, isLoading } = useFetchAlbum(albumId);
 
   if (isLoading) {
@@ -20,7 +21,8 @@ const AlbumPage = ({ albumId }: AlbumPageProps) => {
   }
 
   if (error) {
-    return <ErrorComponent message={error.message} />;
+    const errorMessage = handlePageError(error);
+    return <ErrorComponent message={errorMessage} />;
   }
 
   if (!data) {
@@ -29,6 +31,7 @@ const AlbumPage = ({ albumId }: AlbumPageProps) => {
 
   return (
     <div className="max-w-4xl mobile:mx-6 tablet:mx-6 mx-auto p-6 mt-6 bg-zinc-800 rounded-lg shadow-md">
+      <AlbumSEO album={data} albumId={albumId} />
       <AlbumSection album={data} />
     </div>
   );
